+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+### 2019-05-13
+
+* Updated output to include the correct port numbers based on the chosen template.
+* Added svelte-custom-element template from [https://github.com/gojutin/svelte-custom-element](https://github.com/gojutin/svelte-custom-element)
+
 ## 0.0.1 
 ### 2019-05-07